network: create net_proto files
Move protocol handlers functions into net_proto_handlers.go file and
standardize DNS function names.
